backend/pyproject.toml
[tool.poetry]
name = "decision support api"
version = "0.1.0"
description = "rest api module of the decision support system for the selection of support measures for startups"
authors = ["Vdmrl <vvluzin2004@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.27"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
aiomysql = "^0.2.0"
pymysql = "^1.1.0"
alembic = "^1.13.1"
numpy = "^1.26.4"
sentence-transformers = "^2.6.1"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
gunicorn = "^21.2.0"
cryptography = "^42.0.5"
fastapi-cache2 = { extras = ["redis"], version = "^0.2.1" }
black = "^24.4.2"
autoflake = "^2.3.1"
isort = "^5.13.2"
pyright = "^1.1.370"
flake8-bugbear = "^24.4.26"
sentry-sdk = {extras = ["fastapi"], version = "^2.8.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"